refactor(stock): extract components and add context
